+++ tests/testthat/test_6_qm_summarize.R
@@ -57,9 +57,9 @@ expect_error(qm_summarize(ref = test_sf, key = TRACTCE, category = "postitive",
              "A data set containing map clusters must be specified.")
 
 # test non qm_cluster object input
-# expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = test_tbl, category = "positive",
-#                          count = "clusters"),
-#             "The object test_tbl is not class qm_cluster. The cluster object should be created with qm_combine().", fixed = TRUE)
+expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = test_tbl, category = "positive",
+                          count = "clusters"),
+             "The object test_tbl is not a cluster object. Create a cluster object with qm_combine().", fixed = TRUE)
 
 # test missing category parameter
 expect_error(qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, count = "clusters"),
@@ -101,14 +101,11 @@ resultV1b <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, cat
 
 nrowV1 <- 106
 posV1 <- 0.05660377
-classListV1 <- class(resultV1)
-classListElementV1 <- classListV1[1]
 
 test_that("result object 1 has expected characteristics", {
   expect_equal(nrowV1, nrow(resultV1))
   expect_equal(posV1, mean(resultV1$positive))
-  expect_equal("sf", classListElementV1)
-  expect_equal(resultV1, resultV1b)
+  expect_equal("sf" %in% class(resultV1), TRUE)
 })
 
 resultV2 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive",
@@ -116,13 +113,11 @@ resultV2 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, cate
 
 nrowV2 <- 106
 posV2 <- 1.2
-classListV2 <- class(resultV2)
-classListElementV2 <- classListV2[1]
 
 test_that("result object 2 has expected characteristics", {
   expect_equal(nrowV2, nrow(resultV2))
   expect_equal(posV2, mean(resultV2$positive, na.rm = TRUE))
-  expect_equal("sf", classListElementV2)
+  expect_equal("sf" %in% class(resultV2), TRUE)
 })
 
 resultV3 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, category = "positive",
@@ -130,11 +125,9 @@ resultV3 <- qm_summarize(ref = test_sf, key = TRACTCE, clusters = clusters, cate
 
 nrowV3 <- 5
 posV3 <- 1.2
-classListV3 <- class(resultV3)
-classListElementV3 <- classListV3[1]
 
 test_that("result object 3 has expected characteristics", {
   expect_equal(nrowV3, nrow(resultV3))
   expect_equal(posV3, mean(resultV3$positive))
-  expect_equal("tbl_df", classListElementV3)
+  expect_equal("sf" %in% class(resultV3), FALSE)
 })